[BedTweaks] Update to 1.6
Basically redoes everything. The previous method - patching the beds when they were rendered - does not work anymore. Instead, this changes the data of the beds at the outset using content manager. It generates new bed textures when the game is loaded and modifies Data/Furniture to point to the new textures. Also, removes the ability to change sheet opacity; this will be re-added in the future.