[PlannedParenthood] Remove Harmony Patchall
Just changes Harmony patches from harmony.patchAll (with annotations) to harmony.patch. I thought that this was a 1.6 compatibility update, but when I went to make this commit I realized this was the only change I made; the mod wasn't even broken - I got my old copy and it ran fine. However, the wiki says using patch is better, so since I've done it I might as well push it. This update is not necessary.