Update to 1.6
This is a partial update. It fixes swimming, but does not fix diving.
Updates JsonAssets api to get strings, and the item ids to be strings.
Fixes ambiguous GameLocation.performTouchAction harmony patch, and adds new parameter to GameLocation.isCollidingPosition patch.
Changes the scuba tank to be loaded with Content Manager rather than json assets (json assets isn't working well with shirts rn)
Moves in water but not swimming and on land but swimming checks earlier in the update tick so they don't get skipped by a return statement.
Adds a harmony patch to Game1.warpFarmer to prevent the farmer from being stuck in terrain after going down out of the mountain lake.