Overhaul manual swim logic
Previously, when using manual swim, the game would only check in the direction the player was facing, and if they were clicking, would only do that if they were clicking on the opposite of their current terrain i.e. if they were in water, would only attempt to leave if they were clicking on land.
Now when using manual swim, it checks the direction in which the player is clicking (or rather, the direction of the mouse relative to the player) and then the direction in which they're facing (in that order). Also, adds a config option to allow the player to choose whether they should be required to click on the opposite terrain.