Update conditions for jumping
Now also checks if the player can move before jumping to prevent them jumping when using a tool. This obsoletes all of the old checks for if they were holding a pan, fishing rod..., and removes a bug where if the player was using any tool that wasn't being checked for (like a pickaxe), they would get stuck in an infinite loop.
Additionally, this makes a specific function to preform all of the checks that it is safe for the player to jump. This encompasses all the checks for the old ModEntry.myButtonDown, so that has been removed, plus a couple extras to clean up the UpdateTicked function. To this end, the update for the player's oxygen level was moved to a function in SwimUtils (still called at the same place in UpdateTicked).
Finally, tries to jump in the direction the player is facing before the direction of their cursor (was the other way), and only tries to jump in the direction of their cursor if the manual jump button is a mouse button.