Load assets with content manager
Based on a bug report, it seems that the fact the textures for the tvs were being loaded when the saves were loaded was a problem for some systems - they got an error on mod entry in the spriteBatch.Draw patch because the textures were null. By moving the asset loading to onAssetRequested, the asset will be available whenever needed.
Also fixes the manifest.json version.