Update pillow redrawing
For ease of maintenance, combines the common parts of the furniture and object patches into one method.
Dynamically loads bed textures based on the item id to accommodate beds whose textures are not on the base furniture sheet.
Adds a new BedData class to manage data like each bed's pillow position and whether to redraw the pillows.
Implements a new pillow drawing algorithm to ensure pillows are evenly spaced (the middle pillows used to be closer to each other than to the outer pillows).
The above algorithm also handles the Deluxe Double Bed's unique pillows.
When using sheet transparency, uses BedData to extend the transparency to a height logical for each bed. (previously, the sheets of some beds like the modern bed would have some parts transparent and some not).