Add check for pool-type locations
In SwimUtils.CanSwimHere, checks if the player's current tile is a pool entrance or a swimsuit change tile, and disables the mod in that location if so. This is to detect mod-introduced bathhouse-like areas (it already checks for the bathhouse directly), which otherwise caused the mod to believe the player was swimming on land when they entered the water and created an infinite loop.