Move items to content patcher
Json assets isn't working well with clothes right now, so adds the scuba gear through content patcher.
Adds a Spanish translation for the scuba items.
Fixes the Abigail projectile texture (used to be a white box). Based on a video I saw of past gameplay, I believe it's supposed to be coal, so that's what I changed it to. I also set the parent sheet index to 3 as a fallback of the item isn't defined for some reason (the parent sheet index is ignored if it is defined). I chose 3 because item 3 in the projectiles texture sheet looks the most like coal.