Update to 1.6
Rewrites SocialPage.drawNPCSlot and NPC.tryToReceiveActiveObject transpilers because of changes to the base codes. Does not change their functionality.
Updates drawGiftAmounts to not take a Friendship as a parameter (I couldn't get it to work with the transpiler). Instead, it gets the Friendship from the SocialPage (same end result).
Adds a check for null friendships in drawGiftAmounts so that unmet characters don't crash the mod.