Update warps and hat drawing
Moves all warp data to the swim-map-content.json class.
Resizes a handful of warps so there aren't 'dead spots' at the corners.
Removes a warp that causes the right of the beach to warp back to the right of the beach.
Allows swimming during the Night Market (players were warping in through the river and getting stuck in the water)
Fixes a bug where wearing the scuba gear while swimming would cause the player's torso to be displayed.
Adds a config option to allow the player to wear their hat while wearing their bathing suit.
To facilitate the above, adds a tranpiler to FarmerRenderer.drawHairAndAccessories.