Fix scuba fin buff
Previously, the scuba fin speed buff was not being applied. Updates the logic to reflect the changes in 1.6.
Updates the scuba fin shoe colors so that they are black, rather than white.
Renames the drawHairAndAccessories transpiler because its name was not the same as the class and method it was modifying.
Makes it so that shooting a projectile in the Prairie King game causes the player to face in the direction they shot.
Fixes an error where the Earth Crystal item id was incorrect in SwimMaps.
In SwimUtils.CanSwimHere, changes Game1.currentLocation to Game1.player.currentLocation because the former caused a null pointer exception on exit to menu.