Add support for dehydrator
Adds support for the new dehydrator item.
Adds config options to allow the player to choose whether to display each item type. Additionally, adds a config option to divide the dehydrator item costs by five to account for the fact that it takes five items as input.
Fixes an error where the text that no fertilizer was necessary and the text that no fertilizer could save the crop would not display.