Fix nullpointer in dressers
Now patches all constructors for ShopMenu so that the filter is always initialized. Previously, dressers called a different constructor than the one we used, so it wasn't getting initialized.
Also adds support for tabs - when the player selects a tab, they will only be able to search from items in that tab. Previously, they could search the whole shop no matter what tab.